feat: rename /testing-bot to /ai-testing-agents, refresh page copy and navigation
- Create /ai-testing-agents as canonical page with updated AI agents framing
- Redirect /testing-bot and /bot to /ai-testing-agents
- Update navbar, footer, and internal component links
- Refresh page copy: self-healing tests, Playwright export, autonomous exploration
- Update CTAs from "Get early access" to "Start free trial"
- Footer: "Solutions: AI Testing" with "AI Testing Agents" and "Playwright AI Agent"

Closes #87
