+++ src/components/legacy/home-page/HomeSolutions.tsx
@@ -35,7 +35,7 @@ export default function HomepageHowItWorks(): JSX.Element {
         <div className="card flex flex-1 p-5 md:p-10 justify-between gap-5 shadow-lg">
           <div className="flex flex-col justify-between h-full gap-2">
             <h3 className="text-lg md:text-xl xl:text-2xl font-bold text-secondary-wopee dark:text-primary-wopee">
-              <Link to="/testing-bot">
+              <Link to="/ai-testing-agents">
                 <span className="flex items-center justify-center gap-1">
                   Autonomous Testing Bot <LinkIcon />
                 </span>
@@ -46,7 +46,7 @@ export default function HomepageHowItWorks(): JSX.Element {
             </p>
             <div className="flex justify-center items-center">
               <Link
-                to="/testing-bot"
+                to="/ai-testing-agents"
                 className="group"
               >
                 <Icon
@@ -75,7 +75,7 @@ export default function HomepageHowItWorks(): JSX.Element {
             <p className="text-sm md:text-lg xl:text-xl text-center">
               It is just that simple!
               <Link
-                to="/testing-bot"
+                to="/ai-testing-agents"
                 className="ml-2"
               >
                 Read more..
